Pass 'settings' object to plugins on initalize() and initializeAsync()
