+++ app/lib/initializePlugins.js
@@ -1,12 +1,24 @@
 import { on, send } from 'lib/rpc'
 import plugins from 'plugins'
+import { settings as pluginSettings } from 'lib/plugins'
+
+export const getSettings = (name) => {
+  const settings = pluginSettings.get(name) || {}
+  if (plugins[name].settings) {
+    // Provide default values if nothing is set by user
+    Object.keys(plugins[name].settings).forEach((key) => {
+      settings[key] = settings[key] || plugins[name].settings[key].defaultValue
+    })
+  }
+  return settings
+}
 
 export const initializePlugin = (name) => {
   const { initialize, initializeAsync } = plugins[name]
   if (initialize) {
     // Foreground plugin initialization
     try {
-      initialize()
+      initialize(getSettings(name))
     } catch (e) {
       console.error(`Failed to initialize plugin: ${name}`, e)
     }
